@@ -14,18 +14,19 @@
     "url": "https://github.com/Mercateo/dwatch.git"
   },
   "dependencies": {
-    "babel-polyfill": "^6.7.4",
+    "babel-polyfill": "^6.9.1",
     "dockerode": "^2.2.10",
-    "flag-icon-css": "^1.3.0",
-    "inversify": "^2.0.0-beta.6",
-    "inversify-binding-decorators": "^1.0.0-alpha.4",
+    "flag-icon-css": "^2.3.0",
+    "inversify": "^2.0.0-beta.8",
+    "inversify-binding-decorators": "^1.0.0-alpha.5",
     "material-design-lite": "^1.1.3",
-    "mobx": "^2.2.2",
-    "mobx-react": "^3.3.0",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0",
+    "mobx": "2.2.2",
+    "mobx-react": "^3.3.1",
+    "profilr": "^0.1.4",
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
     "react-intl": "^2.1.2",
-    "react-router": "^2.4.0",
+    "react-router": "^2.4.1",
     "reflect-metadata": "^0.1.3",
     "source-map-support": "^0.4.0"
   }
