@@ -14,7 +14,7 @@
     "start": "cross-env NODE_ENV=development electron ./app",
     "build": "cross-env NODE_ENV=development webpack --config webpack-dev.config.js",
     "build-release": "cross-env NODE_ENV=production webpack --config webpack-prod.config.js",
-    "dist:all": "npm run test && npm run build-release && build --platform=all --arch=x64",
+    "dist:all": "npm run test && npm run build-release && build -l -o -w --x64",
     "dist": "build",
     "test": "cross-env NODE_ENV=production webpack --config webpack-test.config.js && karma start && remap-istanbul -i app/generated/coverage.json > app/generated/coverage-final.json && node scripts/generate-coverage.js text",
     "coverage-html": "npm run test && node scripts/generate-coverage.js html",
@@ -34,39 +34,40 @@
     "buildResources": "assets"
   },
   "devDependencies": {
-    "babel-core": "^6.7.7",
+    "babel-core": "^6.9.1",
     "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-es2015": "^6.9.0",
     "babel-preset-stage-0": "^6.5.0",
     "coveralls": "^2.11.9",
-    "cross-env": "^1.0.7",
+    "cross-env": "^1.0.8",
     "css-loader": "^0.23.1",
-    "electron-builder": "^3.23.0",
-    "electron-prebuilt": "^0.37.6",
-    "emock": "0.0.11",
+    "electron-builder": "^4.2.6",
+    "electron-prebuilt": "^1.2.2",
+    "emock": "^0.1.0",
     "expect": "^1.20.1",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
-    "html-webpack-plugin": "^2.12.0",
-    "inversify-dts": "^1.0.0",
+    "html-webpack-plugin": "^2.19.0",
+    "inversify-dts": "^1.0.5",
     "karma": "^0.13.22",
-    "karma-coverage": "^0.5.5",
-    "karma-electron": "^3.2.0",
-    "karma-mocha": "^0.2.2",
+    "karma-coverage": "^1.0.0",
+    "karma-electron": "^3.2.1",
+    "karma-mocha": "^1.0.1",
     "karma-source-map-support": "^1.1.0",
     "karma-sourcemap-loader": "^0.3.7",
+    "karma-spec-reporter": "0.0.26",
     "mobx-react-devtools": "^4.2.0",
-    "mocha": "^2.4.5",
+    "mocha": "^2.5.3",
     "null-loader": "^0.1.1",
-    "remap-istanbul": "^0.6.3",
-    "style-loader": "^0.13.0",
+    "remap-istanbul": "^0.6.4",
+    "style-loader": "^0.13.1",
     "ts-loader": "^0.8.2",
-    "typescript": "^1.8.10",
-    "typings": "^1.0.3",
+    "typescript": "^1.9.0-dev.20160612-1.0",
+    "typings": "^1.0.4",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.0",
+    "webpack": "^1.13.1",
     "webpack-dev-server": "^1.14.1",
     "webpack-fail-plugin": "^1.0.4",
-    "webpack-node-externals": "^1.0.0"
+    "webpack-node-externals": "^1.2.0"
   }
 }
