@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { ContainerModel } from '../../../models/ContainerModel';
-
-const styles = require('./../../shared/Common.css');
+import { TwoColumnCardRow } from '../../shared/TwoColumnCardRow';
 
 export class NetworkCard extends Component<{container: ContainerModel}, {}> {
   render () {
@@ -16,9 +15,8 @@ export class NetworkCard extends Component<{container: ContainerModel}, {}> {
           </h2>
         </div>
         <div className="mdl-card__supporting-text">
-          <ul className={`${styles.inlineList}`}>
-            <li><FormattedMessage id="containers.th.ports"/></li>
-            <li>
+          <TwoColumnCardRow left={<FormattedMessage id="containers.th.ports"/>}
+                            right={
               <strong>
                 {container.ports.map(port => {
                   let mapping = `${port[ 0 ].port}/${port[ 0 ].protocol}`;
@@ -30,8 +28,7 @@ export class NetworkCard extends Component<{container: ContainerModel}, {}> {
                   return mapping
                 }).join(', ')}
               </strong>
-            </li>
-          </ul>
+              }/>
         </div>
       </div>
     );
