@@ -4,8 +4,7 @@ import { inject } from '../../../utils/IOC';
 import { DockerSystemStore } from '../../../stores/DockerSystemStore';
 import { observer } from 'mobx-react/index';
 import { UiStore } from '../../../stores/UiStore';
-
-const styles = require('./../../shared/Common.css');
+import { TwoColumnCardRow } from '../../shared/TwoColumnCardRow';
 
 @injectIntl
 @observer
@@ -30,83 +29,49 @@ export class SystemCard extends Component<void, {}> {
   render () {
     const { version } = this.dockerSystemStore;
 
-    if (version != null) {
-      return (
-        <div className="mdl-cell mdl-cell--4-col mdl-cell--4-col-tablet mdl-cell--4-col-phone mdl-card mdl-shadow--4dp"
-             style={{ minHeight: '0px' }}>
-          <div className="mdl-card__title">
-            <h2 className="mdl-card__title-text">
-              <FormattedMessage id="home.system"/>
-            </h2>
-          </div>
-          <div className="mdl-card__supporting-text">
-            <ul className={`${styles.inlineList}`}>
-              <li><FormattedMessage id="home.system.server-version"/></li>
-              <li>
-                <strong>
-                  {version.Version}
-                </strong>
-              </li>
-            </ul>
-
-            <ul className={`${styles.inlineList}`}>
-              <li><FormattedMessage id="home.system.os"/></li>
-              <li>
-                <strong>
-                  {version.Os}
-                </strong>
-              </li>
-            </ul>
-
-            <ul className={`${styles.inlineList}`}>
-              <li><FormattedMessage id="home.system.kernel-version"/></li>
-              <li>
-                <strong>
-                  {version.KernelVersion}
-                </strong>
-              </li>
-            </ul>
-
-            <ul className={`${styles.inlineList}`}>
-              <li><FormattedMessage id="home.system.go-version"/></li>
-              <li>
-                <strong>
-                  {version.GoVersion}
-                </strong>
-              </li>
-            </ul>
-
-            <ul className={`${styles.inlineList}`}>
-              <li><FormattedMessage id="home.system.arch"/></li>
-              <li>
-                <strong>
-                  {version.Arch}
-                </strong>
-              </li>
-            </ul>
-
-            <ul className={`${styles.inlineList}`}>
-              <li><FormattedMessage id="home.system.api-version"/></li>
-              <li>
-                <strong>
-                  {version.ApiVersion}
-                </strong>
-              </li>
-            </ul>
-
-            <ul className={`${styles.inlineList}`}>
-              <li><FormattedMessage id="home.system.build-time"/></li>
-              <li>
-                <strong>
-                  <FormattedDate value={version.BuildTime}/>
-                </strong>
-              </li>
-            </ul>
-          </div>
-        </div>
-      );
-    } else {
+    if (version == null) {
       return null;
     }
+
+    return (
+      <div className="mdl-cell mdl-cell--4-col mdl-cell--4-col-tablet mdl-cell--4-col-phone mdl-card mdl-shadow--4dp"
+           style={{ minHeight: '0px' }}>
+        <div className="mdl-card__title">
+          <h2 className="mdl-card__title-text">
+            <FormattedMessage id="home.system"/>
+          </h2>
+        </div>
+        <div className="mdl-card__supporting-text">
+          <TwoColumnCardRow besides={true}
+                            left={
+                               <div>
+                                  <TwoColumnCardRow left={<FormattedMessage id="home.system.server-version"/>}
+                                                    right={<strong>{version.Version}</strong>}/>
+
+                                  <TwoColumnCardRow left={<FormattedMessage id="home.system.os"/>}
+                                                    right={<strong>{version.Os}</strong>}/>
+
+                                  <TwoColumnCardRow left={<FormattedMessage id="home.system.kernel-version"/>}
+                                                    right={<strong>{version.KernelVersion}</strong>}/>
+
+                                  <TwoColumnCardRow left={<FormattedMessage id="home.system.go-version"/>}
+                                                    right={<strong>{version.GoVersion}</strong>}/>
+                               </div>
+                            }
+                            right={
+                               <div>
+                                 <TwoColumnCardRow left={<FormattedMessage id="home.system.arch"/>}
+                                                   right={<strong>{version.Arch}</strong>}/>
+
+                                 <TwoColumnCardRow left={<FormattedMessage id="home.system.api-version"/>}
+                                                   right={<strong>{version.ApiVersion}</strong>}/>
+
+                                 <TwoColumnCardRow left={<FormattedMessage id="home.system.build-time"/>}
+                                                   right={<strong><FormattedDate value={version.BuildTime}/></strong>}/>
+                               </div>
+                            }/>
+        </div>
+      </div>
+    );
   }
 }
