@@ -20,31 +20,7 @@ export class Top extends AutoRefreshComponent<{container: ContainerModel}, {}> {
   private topData: TopModel = null;
 
   render () {
-    if (this.props.container != null && this.props.container.state.runState === CONTAINER_RUN_STATE.RUNNING && this.topData != null) {
-      return (
-        <MDLWrapper>
-          <table
-            className={`mdl-data-table mdl-js-data-table mdl-shadow--2dp ${styles.fullWidthTable} ${styles.fixedLayoutTable}`}>
-            <thead>
-            <tr className="mdl-data-table__cell--non-numeric">
-              {this.topData.Titles != null ? this.topData.Titles.map((title, index) => (
-                <th key={index}>{title}</th>
-              )) : null}
-            </tr>
-            </thead>
-            <tbody>
-            {this.topData.Processes != null ? this.topData.Processes.map((process, index) => (
-              <tr className={`mdl-data-table__cell--non-numeric ${styles.wrap}`} key={index}>
-                {process.map((field, index) => (
-                  <td key={index}>{field}</td>
-                ))}
-              </tr>
-            )) : null}
-            </tbody>
-          </table>
-        </MDLWrapper>
-      );
-    } else {
+    if (this.props.container == null || this.props.container.state.runState !== CONTAINER_RUN_STATE.RUNNING || this.topData == null) {
       return (
         <table
           className={`mdl-data-table mdl-shadow--2dp ${styles.fullWidthTable}`}>
@@ -56,6 +32,30 @@ export class Top extends AutoRefreshComponent<{container: ContainerModel}, {}> {
         </table>
       );
     }
+
+    return (
+      <MDLWrapper>
+        <table
+          className={`mdl-data-table mdl-js-data-table mdl-shadow--2dp ${styles.fullWidthTable} ${styles.fixedLayoutTable}`}>
+          <thead>
+          <tr className="mdl-data-table__cell--non-numeric">
+            {this.topData.Titles != null ? this.topData.Titles.map((title, index) => (
+              <th key={index}>{title}</th>
+            )) : null}
+          </tr>
+          </thead>
+          <tbody>
+          {this.topData.Processes != null ? this.topData.Processes.map((process, index) => (
+            <tr className={`mdl-data-table__cell--non-numeric ${styles.wrap}`} key={index}>
+              {process.map((field, index) => (
+                <td key={index}>{field}</td>
+              ))}
+            </tr>
+          )) : null}
+          </tbody>
+        </table>
+      </MDLWrapper>
+    );
   }
 
   async tick (): Promise<void> {
@@ -63,14 +63,15 @@ export class Top extends AutoRefreshComponent<{container: ContainerModel}, {}> {
   }
 
   private async loadTopData () {
+    if (this.props.container == null || this.props.container.state.runState !== CONTAINER_RUN_STATE.RUNNING) {
+      this.topData = null;
+      return;
+    }
+
     const finishTask = this.uiStore.startAsyncTask();
 
     try {
-      if (this.props.container != null && this.props.container.state.runState === CONTAINER_RUN_STATE.RUNNING) {
-        this.topData = await this.props.container.top();
-      } else {
-        this.topData = null;
-      }
+      this.topData = await this.props.container.top();
       finishTask();
     } catch (e) {
       finishTask(e);
