@@ -3,9 +3,14 @@ import { UiStore } from '../../stores/UiStore';
 import { observer } from 'mobx-react/index';
 import { MDLWrapper } from '../shared/MDLWrapper';
 import { inject } from '../../utils/IOC';
+import { SettingsStore } from '../../stores/SettingsStore';
+import { UpdateNotifications } from './UpdateNoifications';
 
 @observer
 export class Header extends Component<void, {}> {
+  @inject(SettingsStore)
+  private settingsStore: SettingsStore;
+
   @inject(UiStore)
   private uiStore: UiStore;
 
@@ -15,9 +20,11 @@ export class Header extends Component<void, {}> {
         <div className="mdl-layout__header-row">
           <span className="mdl-layout-title">{this.uiStore.pageTitle}</span>
           <div className="mdl-layout-spacer"></div>
-          <div>
-            {this.renderSpinner()}
-          </div>
+
+
+          {this.renderSpinner()}
+          {this.renderUpdateNotifications()}
+
           {/*<div
            className="mdl-textfield mdl-js-textfield mdl-textfield--expandable mdl-textfield--floating-label mdl-textfield--align-right">
            <label className="mdl-button mdl-js-button mdl-button--icon"
@@ -35,14 +42,28 @@ export class Header extends Component<void, {}> {
   }
 
   private renderSpinner () {
-    if (this.uiStore.isLoading) {
-      return (
+    if (!this.uiStore.isLoading) {
+      return null;
+    }
+
+    return (
+      <span>
         <MDLWrapper>
           <div className="mdl-spinner mdl-js-spinner is-active"></div>
         </MDLWrapper>
-      );
-    } else {
+      </span>
+    );
+  }
+
+  private renderUpdateNotifications () {
+    if (!this.settingsStore.showUpdateNotifications) {
       return null;
     }
+
+    return (
+      <span style={{ marginLeft: '16px' }}>
+        <UpdateNotifications/>
+      </span>
+    );
   }
 }
