@@ -22,6 +22,7 @@ module.exports = function(config) {
       'karma-mocha',
       'karma-source-map-support',
       'karma-sourcemap-loader',
+      'karma-spec-reporter',
       'karma-coverage',
       'karma-electron'
     ],
@@ -37,7 +38,7 @@ module.exports = function(config) {
     // test results reporter to use
     // possible values: 'dots', 'progress'
     // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-    reporters: [ 'progress', 'coverage' ],
+    reporters: [ 'spec', 'coverage' ],
 
 
     // web server port
